Fix some menu bar issues and show the file-name in the save dialog window.
Menu bar didnt shrink with ctrl+b or click in the menu's image, now it does
it properly; this was due a lack of send the shrink-bar event for the
shortcut, and adds the event to shrink clicking the image.

When you change the file name in the middle box of the ui it didnt show the
name in the save dialog when save and save as were called, now it show the
name.